Highlighting is working. Click on the highlight toggles the visibility of the note under the highlight.
Highlighting: The notes are all noteHeight in height and even/odd notes
are spaced by lineHeight.  So we need to compute the y for the noteIndex
>> 1, then add the half noteHeight for the odd notes.

Changed the highlighted height to be noteLineHeight (to cover the entire
note).

Beat: Initialize the beat to have all notes on.  Set onClicked to toggle
that state.